Improve cluster details modal
