+++ ui/assets/tweaks.css
@@ -6,6 +6,8 @@
 /* These rules make stacked modals sit nicely */
 .modal-backdrop.show ~ .modal-backdrop.show { z-index: 1080; }
 .modal.show ~ .modal.show { z-index: 2000; }
+.modal-backdrop.show ~ .modal-backdrop.show ~ .modal-backdrop.show { z-index: 2080; }
+.modal.show ~ .modal.show ~ .modal.show { z-index: 3000; }
 
 
 /**
@@ -46,6 +48,22 @@
 }
 
 
+/* Allow an overlay trigger to have the correct cursor */
+.overlay-trigger {
+    cursor: pointer;
+}
+
+
+/* Make progress bars look nice */
+.progress {
+    height: auto;
+    font-size: 0.85rem;
+}
+.progress-bar {
+    padding: 0.25rem 1rem;
+}
+
+
 /**
  * Styles for the notifications container.
  */
@@ -80,16 +98,6 @@
     vertical-align: middle;
     white-space: nowrap;
 }
-.resource-table .progress {
-    height: auto;
-    font-size: 0.85rem;
-}
-.resource-table .progress-bar {
-    padding: 0.25rem 1rem;
-}
-.resource-table .overlay-trigger {
-    cursor: pointer;
-}
 /* The last column of a resource table is always an actions column
    which should shrink to fit the content */
 .resource-table tr > td:last-child {
@@ -116,40 +124,60 @@
 
 
 /**
- * Custom styles for the cluster-type cards.
- */
-.cluster-type-select .card .img-fluid {
-    margin: 0 auto;
-    max-height: 120px;
-}
-.card-body > :last-child {
-    margin-bottom: 0;
-}
-.cluster-parameters-type .img-fluid {
-    max-height: 80px;
-}
-
-
-/**
  * Custom styles for platform cards.
  */
-.platform-card {
+.platform-card, .platform-type-card {
     margin-bottom: 1rem;
 }
-.platform-card > img {
+.platform-card > img, .platform-type-card > img {
     padding: 0.5rem 1rem;
     max-height: 120px;
     object-fit: contain;
 }
-.platform-card > .card-body:not(:first-child) {
-    border-top: 1px solid var(--bs-gray-300) !important;
+.platform-card > .card-body,
+.platform-type-card > .card-body {
     padding: 0.5rem 1rem;
 }
-.platform-card .card-subtitle {
+.platform-card > .card-body:not(:first-child),
+.platform-type-card > .card-body:not(:first-child) {
+    border-top: 1px solid var(--bs-gray-300) !important;
+}
+.platform-card > .list-group + .card-body,
+.platform-type-card > .list-group + .card-body {
+    border-top: none !important;
+}
+.platform-card .card-body > :last-child,
+.platform-type-card .card-body > :last-child {
+    margin-bottom: 0;
+}
+.platform-card .card-subtitle, .platform-type-card .card-subtitle {
     color: var(--bs-gray-600) !important;
 }
 .platform-card .card-header, .platform-card .card-footer {
     text-align: right;
+}
+
+
+/**
+ * Custom styles for the cluster type card.
+ */
+.cluster-type-card {
+    margin-bottom: 1rem;
+}
+.cluster-type-card .row {
+    align-items: stretch;
+}
+.cluster-type-card img {
+    min-width: 70px;
+    max-width: 100%;
+    height: auto;
+    max-height: 100px;
+}
+.cluster-type-card .col > :last-child {
+    margin-bottom: 0;
+}
+.cluster-type-card .card-title {
+    border-bottom: 1px solid var(--bs-gray-300) !important;
 }
